+++ x-pack/plugins/lens/common/expressions/datatable/index.ts
@@ -10,3 +10,5 @@ export * from './datatable';
 export * from './summary';
 export * from './transpose_helpers';
 export * from './utils';
+
+export type { DatatableProps } from './types';
