+++ x-pack/plugins/lens/common/expressions/rename_columns/index.ts
@@ -5,4 +5,4 @@
  * 2.0.
  */
 
-export * from './rename_columns';
+export { renameColumns } from './rename_columns';
