+++ x-pack/plugins/lens/common/expressions/rename_columns/rename_columns.ts
@@ -6,27 +6,9 @@
  */
 
 import { i18n } from '@kbn/i18n';
-import {
-  ExpressionFunctionDefinition,
-  Datatable,
-  DatatableColumn,
-} from '../../../../../../src/plugins/expressions/common';
+import type { RenameColumnsExpressionFunction } from './types';
 
-interface RemapArgs {
-  idMap: string;
-}
-
-type OriginalColumn = { id: string; label: string } & (
-  | { operationType: 'date_histogram'; sourceField: string }
-  | { operationType: string; sourceField: never }
-);
-
-export const renameColumns: ExpressionFunctionDefinition<
-  'lens_rename_columns',
-  Datatable,
-  RemapArgs,
-  Datatable
-> = {
+export const renameColumns: RenameColumnsExpressionFunction = {
   name: 'lens_rename_columns',
   type: 'datatable',
   help: i18n.translate('xpack.lens.functions.renameColumns.help', {
@@ -42,53 +24,9 @@ export const renameColumns: ExpressionFunctionDefinition<
     },
   },
   inputTypes: ['datatable'],
-  fn(data, { idMap: encodedIdMap }) {
-    const idMap = JSON.parse(encodedIdMap) as Record<string, OriginalColumn>;
-
-    return {
-      type: 'datatable',
-      rows: data.rows.map((row) => {
-        const mappedRow: Record<string, unknown> = {};
-        Object.entries(idMap).forEach(([fromId, toId]) => {
-          mappedRow[toId.id] = row[fromId];
-        });
-
-        Object.entries(row).forEach(([id, value]) => {
-          if (id in idMap) {
-            mappedRow[idMap[id].id] = value;
-          } else {
-            mappedRow[id] = value;
-          }
-        });
-
-        return mappedRow;
-      }),
-      columns: data.columns.map((column) => {
-        const mappedItem = idMap[column.id];
-
-        if (!mappedItem) {
-          return column;
-        }
-
-        return {
-          ...column,
-          id: mappedItem.id,
-          name: getColumnName(mappedItem, column),
-        };
-      }),
-    };
+  async fn(...args) {
+    /** Build optimization: prevent adding extra code into initial bundle **/
+    const { renameColumnFn } = await import('./rename_columns_fn');
+    return renameColumnFn(...args);
   },
 };
-
-function getColumnName(originalColumn: OriginalColumn, newColumn: DatatableColumn) {
-  if (originalColumn?.operationType === 'date_histogram') {
-    const fieldName = originalColumn.sourceField;
-
-    // HACK: This is a hack, and introduces some fragility into
-    // column naming. Eventually, this should be calculated and
-    // built more systematically.
-    return newColumn.name.replace(fieldName, originalColumn.label);
-  }
-
-  return originalColumn.label;
-}
