+++ x-pack/plugins/lens/common/expressions/time_scale/time_scale.test.ts
@@ -6,8 +6,8 @@
  */
 
 import moment from 'moment';
-import { Datatable } from 'src/plugins/expressions/public';
-import { TimeRange } from 'src/plugins/data/public';
+import type { Datatable } from 'src/plugins/expressions/public';
+import type { TimeRange } from 'src/plugins/data/public';
 import { functionWrapper } from 'src/plugins/expressions/common/expression_functions/specs/tests/utils';
 
 // mock the specific inner variable:
@@ -22,7 +22,8 @@ jest.mock('../../../../../../src/plugins/data/common/query/timefilter/get_time',
   };
 });
 
-import { getTimeScale, TimeScaleArgs } from './time_scale';
+import { getTimeScale } from './time_scale';
+import type { TimeScaleArgs } from './types';
 
 describe('time_scale', () => {
   let timeScaleWrapped: (input: Datatable, args: TimeScaleArgs) => Promise<Datatable>;
