+++ x-pack/plugins/lens/public/app_plugin/mounter.tsx
@@ -45,7 +45,7 @@ import { getPreloadedState } from '../state_management/lens_slice';
 export async function getLensServices(
   coreStart: CoreStart,
   startDependencies: LensPluginStartDependencies,
-  attributeService: () => Promise<LensAttributeService>
+  attributeService: LensAttributeService
 ): Promise<LensAppServices> {
   const {
     data,
@@ -71,7 +71,7 @@ export async function getLensServices(
     stateTransfer,
     usageCollection,
     savedObjectsTagging,
-    attributeService: await attributeService(),
+    attributeService,
     http: coreStart.http,
     chrome: coreStart.chrome,
     overlays: coreStart.overlays,
@@ -97,8 +97,8 @@ export async function mountApp(
   params: AppMountParameters,
   mountProps: {
     createEditorFrame: EditorFrameStart['createInstance'];
-    attributeService: () => Promise<LensAttributeService>;
-    getPresentationUtilContext: () => Promise<FC>;
+    attributeService: LensAttributeService;
+    getPresentationUtilContext: () => FC;
   }
 ) {
   const { createEditorFrame, attributeService, getPresentationUtilContext } = mountProps;
@@ -252,7 +252,7 @@ export async function mountApp(
 
   params.element.classList.add('lnsAppWrapper');
 
-  const PresentationUtilContext = await getPresentationUtilContext();
+  const PresentationUtilContext = getPresentationUtilContext();
 
   render(
     <I18nProvider>
