+++ x-pack/plugins/lens/public/datatable_visualization/expression.test.tsx
@@ -9,7 +9,7 @@ import type { DatatableProps } from '../../common/expressions';
 import type { LensMultiTable } from '../../common';
 import { createMockExecutionContext } from '../../../../../src/plugins/expressions/common/mocks';
 import type { FormatFactory } from '../../common';
-import { getDatatable } from './expression';
+import { getDatatable } from '../../common/expressions';
 
 function sampleArgs() {
   const indexPatternId = 'indexPatternId';
