+++ x-pack/plugins/lens/public/metric_visualization/metric_suggestions.ts
@@ -6,7 +6,7 @@
  */
 
 import { SuggestionRequest, VisualizationSuggestion, TableSuggestion } from '../types';
-import { MetricState } from '../../common/expressions';
+import type { MetricState } from '../../common/expressions';
 import { layerTypes } from '../../common';
 import { LensIconChartMetric } from '../assets/chart_metric';
