+++ x-pack/plugins/lens/public/pie_visualization/expression.tsx
@@ -19,8 +19,6 @@ import type { FormatFactory } from '../../common';
 import type { PieExpressionProps } from '../../common/expressions';
 import type { ChartsPluginSetup, PaletteRegistry } from '../../../../../src/plugins/charts/public';
 
-export { pie } from '../../common/expressions';
-
 export const getPieRenderer = (dependencies: {
   formatFactory: FormatFactory;
   chartsThemeService: ChartsPluginSetup['theme'];
