+++ x-pack/plugins/lens/public/plugin.ts
@@ -5,28 +5,34 @@
  * 2.0.
  */
 
-import { AppMountParameters, CoreSetup, CoreStart } from 'kibana/public';
+import type { AppMountParameters, CoreSetup, CoreStart } from 'kibana/public';
 import type { Start as InspectorStartContract } from 'src/plugins/inspector/public';
 import type { FieldFormatsSetup, FieldFormatsStart } from 'src/plugins/field_formats/public';
-import { UsageCollectionSetup, UsageCollectionStart } from 'src/plugins/usage_collection/public';
-import { SpacesPluginStart } from '../../spaces/public';
-import { DataPublicPluginSetup, DataPublicPluginStart } from '../../../../src/plugins/data/public';
-import { EmbeddableSetup, EmbeddableStart } from '../../../../src/plugins/embeddable/public';
-import { DashboardStart } from '../../../../src/plugins/dashboard/public';
-import {
+import type {
+  UsageCollectionSetup,
+  UsageCollectionStart,
+} from 'src/plugins/usage_collection/public';
+import type {
+  DataPublicPluginSetup,
+  DataPublicPluginStart,
+} from '../../../../src/plugins/data/public';
+import type { EmbeddableSetup, EmbeddableStart } from '../../../../src/plugins/embeddable/public';
+import type { DashboardStart } from '../../../../src/plugins/dashboard/public';
+import type { SpacesPluginStart } from '../../spaces/public';
+import type {
   ExpressionsServiceSetup,
   ExpressionsSetup,
   ExpressionsStart,
 } from '../../../../src/plugins/expressions/public';
-import {
+import type {
   VisualizationsSetup,
   VisualizationsStart,
 } from '../../../../src/plugins/visualizations/public';
-import { NavigationPublicPluginStart } from '../../../../src/plugins/navigation/public';
-import { UrlForwardingSetup } from '../../../../src/plugins/url_forwarding/public';
-import { GlobalSearchPluginSetup } from '../../global_search/public';
-import { ChartsPluginSetup, ChartsPluginStart } from '../../../../src/plugins/charts/public';
-import { PresentationUtilPluginStart } from '../../../../src/plugins/presentation_util/public';
+import type { NavigationPublicPluginStart } from '../../../../src/plugins/navigation/public';
+import type { UrlForwardingSetup } from '../../../../src/plugins/url_forwarding/public';
+import type { GlobalSearchPluginSetup } from '../../global_search/public';
+import type { ChartsPluginSetup, ChartsPluginStart } from '../../../../src/plugins/charts/public';
+import type { PresentationUtilPluginStart } from '../../../../src/plugins/presentation_util/public';
 import { EmbeddableStateTransfer } from '../../../../src/plugins/embeddable/public';
 import type { EditorFrameService as EditorFrameServiceType } from './editor_frame_service';
 import { IndexPatternFieldEditorStart } from '../../../../src/plugins/index_pattern_field_editor/public';
@@ -51,29 +57,33 @@ import type {
   PieVisualizationPluginSetupPlugins,
 } from './pie_visualization';
 import type { HeatmapVisualization as HeatmapVisualizationType } from './heatmap_visualization';
-import { AppNavLinkStatus } from '../../../../src/core/public';
 import type { SavedObjectTaggingPluginStart } from '../../saved_objects_tagging/public';
+
+import { AppNavLinkStatus } from '../../../../src/core/public';
 
 import {
   UiActionsStart,
   ACTION_VISUALIZE_FIELD,
   VISUALIZE_FIELD_TRIGGER,
 } from '../../../../src/plugins/ui_actions/public';
-import { APP_ID, FormatFactory, getEditPath, NOT_INTERNATIONALIZED_PRODUCT_NAME } from '../common';
-import type { EditorFrameStart, VisualizationType } from './types';
+import { APP_ID, getEditPath, NOT_INTERNATIONALIZED_PRODUCT_NAME } from '../common/constants';
+import type { FormatFactory } from '../common/types';
+import type { VisualizationType } from './types';
 import { getLensAliasConfig } from './vis_type_alias';
 import { visualizeFieldAction } from './trigger_actions/visualize_field_actions';
-import { getSearchProvider } from './search_provider';
 
-import { LensAttributeService } from './lens_attribute_service';
-import { LensEmbeddableInput } from './embeddable';
+import type { LensEmbeddableInput } from './embeddable';
 import { EmbeddableFactory, LensEmbeddableStartServices } from './embeddable/embeddable_factory';
 import {
   EmbeddableComponentProps,
   getEmbeddableComponent,
 } from './embeddable/embeddable_component';
 import { getSaveModalComponent } from './app_plugin/shared/saved_modal_lazy';
-import { SaveModalContainerProps } from './app_plugin/save_modal_container';
+import type { SaveModalContainerProps } from './app_plugin/save_modal_container';
+
+import { createStartServicesGetter } from '../../../../src/plugins/kibana_utils/public';
+import { setupExpressions } from './expressions';
+import { getSearchProvider } from './search_provider';
 
 export interface LensPluginSetupDependencies {
   urlForwarding: UrlForwardingSetup;
@@ -154,8 +164,6 @@ export interface LensPublicStart {
 export class LensPlugin {
   private datatableVisualization: DatatableVisualizationType | undefined;
   private editorFrameService: EditorFrameServiceType | undefined;
-  private createEditorFrame: EditorFrameStart['createInstance'] | null = null;
-  private attributeService: (() => Promise<LensAttributeService>) | null = null;
   private indexpatternDatasource: IndexPatternDatasourceType | undefined;
   private xyVisualization: XyVisualizationType | undefined;
   private metricVisualization: MetricVisualizationType | undefined;
@@ -178,37 +186,32 @@ export class LensPlugin {
       usageCollection,
     }: LensPluginSetupDependencies
   ) {
-    this.attributeService = async () => {
-      const { getLensAttributeService } = await import('./async_services');
-      const [coreStart, startDependencies] = await core.getStartServices();
-      return getLensAttributeService(coreStart, startDependencies);
-    };
+    const startServices = createStartServicesGetter(core.getStartServices);
 
     const getStartServices = async (): Promise<LensEmbeddableStartServices> => {
-      const [coreStart, deps] = await core.getStartServices();
+      const { getLensAttributeService } = await import('./async_services');
+      const { core: coreStart, plugins } = startServices();
 
-      this.initParts(
+      await this.initParts(
         core,
         data,
-        embeddable,
         charts,
         expressions,
-        usageCollection,
         fieldFormats,
-        deps.fieldFormats.deserialize
+        plugins.fieldFormats.deserialize
       );
 
       return {
-        attributeService: await this.attributeService!(),
+        attributeService: getLensAttributeService(coreStart, plugins),
         capabilities: coreStart.application.capabilities,
         coreHttp: coreStart.http,
-        timefilter: deps.data.query.timefilter.timefilter,
-        expressionRenderer: deps.expressions.ReactExpressionRenderer,
+        timefilter: plugins.data.query.timefilter.timefilter,
+        expressionRenderer: plugins.expressions.ReactExpressionRenderer,
         documentToExpression: this.editorFrameService!.documentToExpression,
-        indexPatternService: deps.data.indexPatterns,
-        uiActions: deps.uiActions,
+        indexPatternService: plugins.data.indexPatterns,
+        uiActions: plugins.uiActions,
         usageCollection,
-        inspector: deps.inspector,
+        inspector: plugins.inspector,
       };
     };
 
@@ -218,17 +221,22 @@ export class LensPlugin {
 
     visualizations.registerAlias(getLensAliasConfig());
 
-    const getPresentationUtilContext = async () => {
-      const [, deps] = await core.getStartServices();
-      const { ContextProvider } = deps.presentationUtil;
-      return ContextProvider;
-    };
+    setupExpressions(
+      expressions,
+      () => startServices().plugins.fieldFormats.deserialize,
+      async () => {
+        const { getTimeZone } = await import('./utils');
+        return getTimeZone(core.uiSettings);
+      }
+    );
+
+    const getPresentationUtilContext = () =>
+      startServices().plugins.presentationUtil.ContextProvider;
 
     const ensureDefaultIndexPattern = async () => {
-      const [, deps] = await core.getStartServices();
       // make sure a default index pattern exists
       // if not, the page will be redirected to management and visualize won't be rendered
-      await deps.data.indexPatterns.ensureDefaultIndexPattern();
+      await startServices().plugins.data.indexPatterns.ensureDefaultIndexPattern();
     };
 
     core.application.register({
@@ -236,25 +244,27 @@ export class LensPlugin {
       title: NOT_INTERNATIONALIZED_PRODUCT_NAME,
       navLinkStatus: AppNavLinkStatus.hidden,
       mount: async (params: AppMountParameters) => {
-        const [, deps] = await core.getStartServices();
+        const { core: coreStart, plugins: deps } = startServices();
 
         await this.initParts(
           core,
           data,
-          embeddable,
           charts,
           expressions,
-          usageCollection,
           fieldFormats,
           deps.fieldFormats.deserialize
         );
 
-        const { mountApp, stopReportManager } = await import('./async_services');
+        const { mountApp, stopReportManager, getLensAttributeService } = await import(
+          './async_services'
+        );
+        const frameStart = this.editorFrameService!.start(coreStart, deps);
+
         this.stopReportManager = stopReportManager;
         await ensureDefaultIndexPattern();
         return mountApp(core, params, {
-          createEditorFrame: this.createEditorFrame!,
-          attributeService: this.attributeService!,
+          createEditorFrame: frameStart.createInstance,
+          attributeService: getLensAttributeService(coreStart, deps),
           getPresentationUtilContext,
         });
       },
@@ -280,10 +290,8 @@ export class LensPlugin {
   private async initParts(
     core: CoreSetup<LensPluginStartDependencies, void>,
     data: DataPublicPluginSetup,
-    embeddable: EmbeddableSetup | undefined,
     charts: ChartsPluginSetup,
     expressions: ExpressionsServiceSetup,
-    usageCollection: UsageCollectionSetup | undefined,
     fieldFormats: FieldFormatsSetup,
     formatFactory: FormatFactory
   ) {
@@ -303,13 +311,9 @@ export class LensPlugin {
     this.metricVisualization = new MetricVisualization();
     this.pieVisualization = new PieVisualization();
     this.heatmapVisualization = new HeatmapVisualization();
-    const editorFrameSetupInterface = this.editorFrameService.setup(core, {
-      data,
-      embeddable,
-      charts,
-      expressions,
-      usageCollection,
-    });
+
+    const editorFrameSetupInterface = this.editorFrameService.setup();
+
     const dependencies: IndexPatternDatasourceSetupPlugins &
       XyVisualizationPluginSetupPlugins &
       DatatableVisualizationPluginSetupPlugins &
@@ -328,9 +332,6 @@ export class LensPlugin {
     this.metricVisualization.setup(core, dependencies);
     this.pieVisualization.setup(core, dependencies);
     this.heatmapVisualization.setup(core, dependencies);
-    const [coreStart, startDependencies] = await core.getStartServices();
-    const frameStart = this.editorFrameService.start(coreStart, startDependencies);
-    this.createEditorFrame = frameStart.createInstance;
   }
 
   start(core: CoreStart, startDependencies: LensPluginStartDependencies): LensPublicStart {
@@ -345,7 +346,7 @@ export class LensPlugin {
 
     return {
       EmbeddableComponent: getEmbeddableComponent(core, startDependencies),
-      SaveModalComponent: getSaveModalComponent(core, startDependencies, this.attributeService!),
+      SaveModalComponent: getSaveModalComponent(core, startDependencies),
       navigateToPrefilledEditor: (
         input,
         { openInNewTab = false, originatingApp = '', originatingPath } = {}
