+++ x-pack/plugins/lens/public/search_provider.ts
@@ -5,12 +5,13 @@
  * 2.0.
  */
 
-import { ApplicationStart } from 'kibana/public';
+import type { ApplicationStart } from 'kibana/public';
 import { from, of } from 'rxjs';
 import { i18n } from '@kbn/i18n';
 import { DEFAULT_APP_CATEGORIES } from '../../../../src/core/public';
-import { GlobalSearchResultProvider } from '../../global_search/public';
-import { getFullPath } from '../common';
+import { getFullPath } from '../common/constants';
+
+import type { GlobalSearchResultProvider } from '../../global_search/public';
 
 /**
  * Global search provider adding a Lens entry.
