+++ x-pack/plugins/lens/public/vis_type_alias.ts
@@ -6,8 +6,8 @@
  */
 
 import { i18n } from '@kbn/i18n';
-import { VisTypeAlias } from 'src/plugins/visualizations/public';
-import { getBasePath, getEditPath } from '../common';
+import type { VisTypeAlias } from 'src/plugins/visualizations/public';
+import { getBasePath, getEditPath } from '../common/constants';
 
 export const getLensAliasConfig = (): VisTypeAlias => ({
   aliasPath: getBasePath(),
